+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -115,4 +115,27 @@ final class ArrayExtensionsTests: XCTestCase {
         ]
         XCTAssertEqual(arrayWithoutDuplicatesNHashable, arrayWithoutDuplicatesNHashablePrepared)
     }
+    
+    func testAppendElement() {
+        var testArray = ["h", "e", "l", "l", "o"]
+        let optionalString: String? = "f"
+        testArray.appendIfNonNil(optionalString)
+        XCTAssertEqual(testArray, ["h", "e", "l", "l", "o", "f"])
+        
+        let nilString: String? = nil
+        testArray.appendIfNonNil(nilString)
+        XCTAssertEqual(testArray, ["h", "e", "l", "l", "o", "f"])
+    }
+    
+    func testAppendSequence() {
+        var testEmptyArray: [Double] = []
+        let numbersToAppend: [Double]? = [69.0, 68.0, 67.0]
+        testEmptyArray.appendIfNonNil(contentsOf: numbersToAppend)
+        XCTAssertEqual(testEmptyArray, numbersToAppend)
+        
+        var testStringArray: [String] = ["h", "e", "l", "l", "o"]
+        let nilArray: [String]? = nil
+        testStringArray.appendIfNonNil(contentsOf: nilArray)
+        XCTAssertEqual(testStringArray, ["h", "e", "l", "l", "o"])
+    }
 }